comments

+++ a_player/views.py
@@ -230,9 +230,9 @@ def ajax_del_comment(request):
     imdb_id = request.POST.get('imdb_id')
     comment_id = request.POST.get('commentId')
     film = FilmModel.objects.filter(imdb_id=imdb_id)
-    # user = CommentModel.objects.filter(pk=comment_id, user=request.user)
-    # if len(user) == 0:
-    #     return HttpResponse("error4")
+    user = CommentModel.objects.filter(pk=comment_id, user=request.user)
+    if len(user) == 0:
+        return HttpResponse("error4")
     if len(film) == 0:
         return HttpResponse("error5")
     CommentModel.objects.filter(pk=comment_id).delete()
